requirements.txt
pdf2image==1.17.0
opencv-python==4.8.1.78
pytesseract==0.3.10
PyPDF2==3.0.1
img2table==1.3.0
streamlit==1.28.0
Pillow==10.0.1
PyMuPDF==1.23.0
img2table[paddle]
pandas==2.1.3
openpyxl==3.1.2
pyyaml==6.0.1
pytest==7.4.3
pytest-cov==4.1.0
black==23.10.1
flake8==6.1.0
mypy==1.7.0
types-PyYAML==6.0.12.12